feat: 인기상품 조회 api 주문 수량 반환 값 타입 수정

+++ src/main/java/com/study/ecommerce/domain/product/ProductInfo.java
@@ -13,7 +13,7 @@ public class ProductInfo {
 		Long productId,
 		String productName,
 		Long productPrice,
-		Long orderAmount
+		Integer orderAmount
 	) {
 	}
